Fixed splitting fasta header by space or tab
In python, the code `if (" " or "\t") in line:` evaluates to `if " " in line:` since " " is a non-empty string.
What you really wanted to do is check like this: `if " " in line or "\t" in line:` to check both, but this introduces
another problem that you need to then check which one was in the line to be able to chop everything off past that.

This is all solved using regex as I have modified to fix the bug.